Update project.json to include the .NET Core 1.1 packages
Switched to using Microsoft.NETCore.App meta-package and removed
some redundant package references that were part of that closure.

Manuall excluded Libuv, Microsoft.CodeAnalysis.* which are mostly about
ASP.Net scenarios which aren't likely interesting for server.

Worked around a missing Jit package for arm by including a reference to a
newer beta package and then excluding it because it isn't actually needed
it is a broken runtime package dependency.